// src/app/shared/generated/model/toolPageResult.ts
/**
 * onecx-ai-provider  service
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Tool } from './tool';


export interface ToolPageResult { 
    totalElements?: number;
    number?: number;
    size?: number;
    totalPages?: number;
    stream?: Array<Tool>;
}